Added utils module, containing retry requests, write to file, build transaction data, and command line argument splitter functions Changed definition of retry requests function to accept generic arguments Added initialization of nonce and used retry requests function to check if subgraph has updated data at the start of every loop iteration Moved cluster snapshot formatting to the function that gets the cluster snapshot itself
